google-admin-sdk-api/requirements.txt
google-api-python-client==1.6.6
google-auth==1.4.1
google-auth-httplib2==0.0.3